move updateComplete to test helpers
