+++ packages/dashboard/test/helpers.ts
@@ -1,4 +1,5 @@
 import { expect } from '@vaadin/chai-plugins';
+import { nextFrame, nextUpdate } from '@vaadin/testing-helpers';
 import sinon from 'sinon';
 import type { DashboardSection } from '../src/vaadin-dashboard-section.js';
 import type { DashboardWidget } from '../src/vaadin-dashboard-widget.js';
@@ -373,4 +374,13 @@ function onceInvoked(object, functionName): Promise<void> {
 
 export async function onceResized(dashboard: HTMLElement): Promise<void> {
   await onceInvoked(dashboard, '_onResize');
+}
+
+export async function updateComplete(dashboard: HTMLElement): Promise<void> {
+  await nextUpdate(dashboard);
+  const widgets = dashboard.querySelectorAll('vaadin-dashboard-widget');
+  for (const widget of widgets) {
+    await nextUpdate(widget);
+  }
+  await nextFrame();
 }
